test(review.t): do not use `which python`
It fails on Windows:

$ virtualenv --python="$(which python)" FOO
FileNotFoundError: [WinError 2] The system cannot find the file specified: 'C:/Users/a/AppData/Local/Programs/Python/Python310/python'

I believe it is because of missing .exe extension of the file.

I think virtualenv uses the first found python by default
so this should not change anything.